Fixed issues

+++ app/point.py
@@ -11,8 +11,8 @@ class Point:
             return False
         return self.x == other.x and self.y == other.y
 
-    def __hash__(self) -> float:
-        return self._x + self._y
+    def __hash__(self) -> int:
+        return int(self._x + self._y)
 
     @property
     def x(self) -> float:
